Fenced frames: Event reporting support for ad component urn iframes.
Modify GetFencedFrameProperties to take in a boolean argument, which
allows clients to obtain the fenced frame properties associated with
the frame itself. This is useful when the frame invokes event reporting
APIs as the reporting metadata should be retrieved from the frame
itself. Without this change, ad component urn iframes nested in the
main ad fenced frame will use the fenced frame's FencedFrameProperties
for event reporting, which is incorrect.

This CL also parameterized the test case so that it tests the
following cases:
1. Ad in fenced frame, ad component in a nested fenced frame.
2. Ad in fenced frame, ad component in a nested iframe.
3. Ad in iframe, ad component in a nested fenced frame.
4. Ad in iframe, ad component in a nested iframe.

Bug: 1396067
Change-Id: I56fba3a93f9ce9c4ffbdac5074e3ef701ac25967